+++ app.py
@@ -4,7 +4,8 @@ from flask import Flask
 
 from mlrvalidator.site_file_validator_rules import SitefileValidator
 from mlrvalidator.site_file_validator_warnings import SitefileWarningValidator
-from mlrvalidator.schema import error_schema, warning_schema
+from mlrvalidator.site_file_cross_field_validator_rules import CrossFieldValidator
+from mlrvalidator.schema import error_schema, warning_schema, cross_field_schema
 
 
 application = Flask(__name__)
@@ -20,6 +21,9 @@ sitefile_error_validator.allow_unknown = True
 
 sitefile_warning_validator = SitefileWarningValidator(warning_schema)
 sitefile_warning_validator.allow_unknown = True
+
+sitefile_crossfield_error_validator = CrossFieldValidator(cross_field_schema)
+sitefile_crossfield_error_validator.allow_unknown = True
 
 from mlrvalidator.services import *
