+++ mlrvalidator/schema.py
@@ -10,6 +10,10 @@ fd = open(os.path.join(PROJECT_DIR, 'schemas/error_schema.yml'))
 with fd:
     error_schema = yaml.load(fd.read())
 
+fd = open(os.path.join(PROJECT_DIR, 'schemas/cross_field_schema.yml'))
+with fd:
+    cross_field_schema = yaml.load(fd.read())
+
 fdw = open(os.path.join(PROJECT_DIR, 'schemas/warning_schema.yml'))
 with fdw:
     warning_schema = yaml.load(fdw.read())
